Fix Windows warnings
Submitted on behalf of Patrick Hodoul, with some minor mods by LG.

+++ src/include/OpenImageIO/ustring.h
@@ -834,7 +834,7 @@ iequals(const std::string& a, ustring b)
 // ustring variant stof from OpenImageIO/strutil.h
 namespace Strutil {
 
-inline int
+inline float
 stof(ustring s)
 {
     return Strutil::stof(s.string());
